backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
pytest-django>=4.7
pytest>=7.4
factory-boy>=3.3
django-prometheus>=2.3
django-ratelimit>=4.1
